fixed landing page and study group page css
