+++ src/ui-components/SLAnnouncementCollection.jsx
@@ -6,14 +6,14 @@
 
 /* eslint-disable */
 import * as React from "react";
-import { listStudyGroupCards } from "../graphql/queries";
-import StudyGroupCard from "./StudygroupCard";
+import { listAnnouncements } from "../graphql/queries";
+import SocialPost from "./SocialPost";
 import { getOverrideProps } from "./utils";
 import { Collection, Pagination, Placeholder } from "@aws-amplify/ui-react";
 import { API } from "aws-amplify";
 const nextToken = {};
 const apiCache = {};
-export default function StudyGroupCardCollection(props) {
+export default function AnnouncementCollection(props) {
   const { items: itemsProp, overrideItems, overrides, ...rest } = props;
   const [pageIndex, setPageIndex] = React.useState(1);
   const [hasMorePages, setHasMorePages] = React.useState(true);
@@ -22,8 +22,8 @@ export default function StudyGroupCardCollection(props) {
   const [instanceKey, setInstanceKey] = React.useState("newGuid");
   const [loading, setLoading] = React.useState(true);
   const [maxViewed, setMaxViewed] = React.useState(1);
-  const pageSize = 6;
-  const isPaginated = false;
+  const pageSize = 2;
+  const isPaginated = true;
   React.useEffect(() => {
     nextToken[instanceKey] = "";
     apiCache[instanceKey] = [];
@@ -40,7 +40,11 @@ export default function StudyGroupCardCollection(props) {
   const jumpToPage = (pageNum) => {
     setPageIndex(pageNum);
   };
-  const loadPage = async (page) => {
+  React.useEffect(() => {
+    loadPage(pageIndex, props.studyGroupId); // Pass studyGroupId prop
+  }, [pageIndex, props.studyGroupId]);
+
+  const loadPage = async (page, studyGroupId) => {
     const cacheUntil = page * pageSize + 1;
     const newCache = apiCache[instanceKey].slice();
     let newNext = nextToken[instanceKey];
@@ -48,16 +52,17 @@ export default function StudyGroupCardCollection(props) {
       setLoading(true);
       const variables = {
         limit: pageSize,
+        filter: { studygroupcardID: { eq: studyGroupId } },
       };
       if (newNext) {
         variables["nextToken"] = newNext;
       }
       const result = (
         await API.graphql({
-          query: listStudyGroupCards.replaceAll("__typename", ""),
+          query: listAnnouncements.replaceAll("__typename", ""),
           variables,
         })
-      ).data.listStudyGroupCards;
+      ).data.listAnnouncements;
       newCache.push(...result.items);
       newNext = result.nextToken;
     }
@@ -71,24 +76,24 @@ export default function StudyGroupCardCollection(props) {
     nextToken[instanceKey] = newNext;
   };
   React.useEffect(() => {
-    loadPage(pageIndex);
+    loadPage(pageIndex, props.studyGroupID);
   }, [pageIndex]);
+  
   React.useEffect(() => {
     setMaxViewed(Math.max(maxViewed, pageIndex));
   }, [pageIndex, maxViewed, setMaxViewed]);
   return (
     <div>
       <Collection
-        type="grid"
+        type="list"
+        isSearchable="true"
         searchPlaceholder="Search..."
-        templateColumns="1fr 1fr 1fr"
-        autoFlow="row"
-        alignItems="stretch"
-        justifyContent="stretch"
+        direction="column"
+        justifyContent="center"
         itemsPerPage={pageSize}
         isPaginated={!isApiPagination && isPaginated}
         items={itemsProp || (loading ? new Array(pageSize).fill({}) : items)}
-        {...getOverrideProps(overrides, "StudyGroupCardCollection")}
+        {...getOverrideProps(overrides, "AnnouncementCollection")}
         {...rest}
       >
         {(item, index) => {
@@ -96,12 +101,12 @@ export default function StudyGroupCardCollection(props) {
             return <Placeholder key={index} size="large" />;
           }
           return (
-            <StudyGroupCard
-              studyGroupCard={item}
-              margin="10px 10px 10px 10px"
+            <SocialPost
+              announcement={item}
+              margin="0 0 5px 0"
               key={item.id}
               {...(overrideItems && overrideItems({ item, index }))}
-            ></StudyGroupCard>
+            ></SocialPost>
           );
         }}
       </Collection>
