+++ src/ui-components/StandardCardCollectionB.jsx
@@ -6,14 +6,14 @@
 
 /* eslint-disable */
 import * as React from "react";
-import { listAnnouncements } from "../graphql/queries";
-import SocialPost from "./SocialPost";
+import { listStudyGroupCards } from "../graphql/queries";
+import StandardCard from "./StandardCard";
 import { getOverrideProps } from "./utils";
 import { Collection, Pagination, Placeholder } from "@aws-amplify/ui-react";
 import { API } from "aws-amplify";
 const nextToken = {};
 const apiCache = {};
-export default function AnnouncementCollection(props) {
+export default function StandardCardCollectionB(props) {
   const { items: itemsProp, overrideItems, overrides, ...rest } = props;
   const [pageIndex, setPageIndex] = React.useState(1);
   const [hasMorePages, setHasMorePages] = React.useState(true);
@@ -40,12 +40,7 @@ export default function AnnouncementCollection(props) {
   const jumpToPage = (pageNum) => {
     setPageIndex(pageNum);
   };
-  
-  React.useEffect(() => {
-    loadPage(pageIndex, props.studyGroupId); // Pass studyGroupId prop
-  }, [pageIndex, props.studyGroupId]);
-
-  const loadPage = async (page,studyGroupId) => {
+  const loadPage = async (page) => {
     const cacheUntil = page * pageSize + 1;
     const newCache = apiCache[instanceKey].slice();
     let newNext = nextToken[instanceKey];
@@ -53,18 +48,17 @@ export default function AnnouncementCollection(props) {
       setLoading(true);
       const variables = {
         limit: pageSize,
-        filter: { studygroupcardID: { eq: studyGroupId } },
+        filter: { acceptingMembers: { eq: true } },
       };
-      console.log(variables);
       if (newNext) {
         variables["nextToken"] = newNext;
       }
       const result = (
         await API.graphql({
-          query: listAnnouncements.replaceAll("__typename", ""),
+          query: listStudyGroupCards.replaceAll("__typename", ""),
           variables,
         })
-      ).data.listAnnouncements;
+      ).data.listStudyGroupCards;
       newCache.push(...result.items);
       newNext = result.nextToken;
     }
@@ -78,7 +72,7 @@ export default function AnnouncementCollection(props) {
     nextToken[instanceKey] = newNext;
   };
   React.useEffect(() => {
-    loadPage(pageIndex, props.studyGroupId);
+    loadPage(pageIndex);
   }, [pageIndex]);
   React.useEffect(() => {
     setMaxViewed(Math.max(maxViewed, pageIndex));
@@ -86,15 +80,17 @@ export default function AnnouncementCollection(props) {
   return (
     <div>
       <Collection
-        type="list"
-        isSearchable="true"
+        type="grid"
+        isSearchable={true}
         searchPlaceholder="Search..."
-        direction="column"
-        justifyContent="center"
+        templateColumns="1fr 1fr 1fr"
+        autoFlow="row"
+        alignItems="stretch"
+        justifyContent="stretch"
         itemsPerPage={pageSize}
         isPaginated={!isApiPagination && isPaginated}
         items={itemsProp || (loading ? new Array(pageSize).fill({}) : items)}
-        {...getOverrideProps(overrides, "AnnouncementCollection")}
+        {...getOverrideProps(overrides, "StandardCardCollectionB")}
         {...rest}
       >
         {(item, index) => {
@@ -102,12 +98,14 @@ export default function AnnouncementCollection(props) {
             return <Placeholder key={index} size="large" />;
           }
           return (
-            <SocialPost
-              announcement={item}
-              margin="0 0 5px 0"
+            <StandardCard
+              studyGroupCard={item}
+              height="auto"
+              width="auto"
+              margin="0 15px 15px 0"
               key={item.id}
               {...(overrideItems && overrideItems({ item, index }))}
-            ></SocialPost>
+            ></StandardCard>
           );
         }}
       </Collection>
